Add support for XDG compliant database file location
I've added XdgPartialConfigProvider providing a stable location for the
main database file. It is still possible to indicate where the db file
should reside via CLI args or env.
