+++ dsync-server/.env
@@ -1,2 +1,2 @@
-DATABASE_URL=./main.db
+# DATABASE_URL=./main.db
 SERVER_PORT=50051
