+++ dsync-server/Cargo.toml
@@ -27,3 +27,4 @@ bytes = "1.10.1"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.145"
 async-trait = "0.1.89"
+xdg = "3.0.0"
