+++ dsync-server/src/bin/dsync-server/main.rs
@@ -23,8 +23,9 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     };
 
     let config = load_config(&[
-        &config::provider::CliArgsConfigProvider::new(args.clone()),
+        &config::provider::CliArgsConfigProvider::new(args),
         &config::provider::EnvConfigProvider::new(),
+        &config::provider::XdgPartialConfigProvider::new(),
     ])
     .context("Failed to load configuration")?;
